Reduce TF training instance scaling
